% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2o_context.R
\name{as_h2o_frame}
\alias{as_h2o_frame}
\title{Convert a Spark DataFrame to an H2O Frame}
\usage{
as_h2o_frame(sc, x, name = NULL)
}
\arguments{
\item{sc}{Object of type \code{spark_connection}.}

\item{x}{A \code{spark_dataframe}.}

\item{name}{The name of the H2OFrame}
}
\description{
Convert a Spark DataFrame to an H2O Frame
}

