% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_context.R
\name{h2o_context}
\alias{h2o_context}
\title{Get the H2OContext from another object}
\usage{
h2o_context(x, ...)
}
\arguments{
\item{x}{Object of type \code{spark_connection} or \code{spark_jobj}}
}
\description{
Get the H2OContext. Will create the context if it hasn't been previously created.
}

