% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_context.R
\name{h2o_flow}
\alias{h2o_flow}
\title{Open the H2O Flow UI in a browser}
\usage{
h2o_flow(sc, strict_version_check = FALSE)
}
\arguments{
\item{sc}{Object of type \code{spark_connection}.}

\item{strict_version_check}{(Optional) Setting this to TRUE performs an strinct version check of H2O.}
}
\description{
Open the H2O Flow UI in a browser
}

