% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_h2o.R
\name{install_h2o}
\alias{install_h2o}
\title{An easy installation of the H2O R pacakage}
\usage{
install_h2o(release_name = "rel-tverberg", release_number = "1")
}
\arguments{
\item{release_name}{Object of type character that specifies the release name of the H2O pacakge}

\item{release_number}{Object of type character that specifies the release number of the H2O pacakge}
}
\description{
An easy installation of the H2O R pacakage
}
\examples{
\donttest{
#Install the latest release of H2O on 1/30/16 (relv-tverberg-1)
install_h2o(release_name = "rel-tverberg", release_number = "1")
}
}

