% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_context.R
\name{as_h2o_frame}
\alias{as_h2o_frame}
\title{Convert a Spark DataFrame to an H2O Frame}
\usage{
as_h2o_frame(x)
}
\arguments{
\item{x}{A \code{spark_dataframe}.}
}
\description{
Convert a Spark DataFrame to an H2O Frame
}

