% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2o_context.R
\name{as_spark_dataframe}
\alias{as_spark_dataframe}
\title{Convert an H2O Frame to a Spark DataFrame}
\usage{
as_spark_dataframe(sc, x)
}
\arguments{
\item{sc}{Object of type \code{spark_connection}.}

\item{x}{An \code{H2OFrame}.}
}
\description{
Convert an H2O Frame to a Spark DataFrame
}

