% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2o_context.R
\name{h2o_flow}
\alias{h2o_flow}
\title{Open the H2O Flow UI in a browser}
\usage{
h2o_flow(sc)
}
\arguments{
\item{sc}{Object of type \code{spark_connection}.}
}
\description{
Open the H2O Flow UI in a browser
}

